+++ api.py
@@ -7,7 +7,7 @@ from resources.grupoDeTrabajoResource import GrupoDeTrabajoID,GrupoDeTrabajo,Gru
 from resources.experimentoResource import ExperimentoResource, ExperimentoMuestra,ObtenerBlogsExp,Experimentos, CerrarExperimento
 from resources.contenedorResource import Contenedor, ContenedorProyecto, ContenedorParent,ContenedorProyectoId
 
-from resources.grupoExperimentalResource import GrupoExperimental, GruposExperimentales, DevidirGrupoExperimental
+from resources.grupoExperimentalResource import GrupoExperimental, GruposExperimentales, DividirGrupoExperimental
 from resources.stockResource import ObtenerProductosStock,ProductoEnStock,BorrarTodoStock,ConsumirStockResource,ProductoEnStockID
 from resources.productoResource import ProductoEnStockDeGrupos,ProductoResource,ObtenerProductosResource,ProductoID,ArchivoProducto
 from resources.distribuidoraResource import DistribuidoraResource,ObtenerDistribuidorasResource,DistribuidoraID
@@ -99,7 +99,8 @@ api.add_resource(ObtenerBlogsExp, '/api/v1/blogExperimento', endpoint='obtener_b
 api.add_resource(GrupoExperimental, '/api/v1/grupoExperimental/<int:idGrupoExperimental>', endpoint='grupo_experimental')
 api.add_resource(GrupoExperimental, '/api/v1/nuevoGrupoExperimental', endpoint='nuevo_grupo_experimental')
 api.add_resource(GruposExperimentales, '/api/v1/experimento/<int:idExperimento>/gruposExperimentales', endpoint='grupos_experimentales_del_experimento')
-api.add_resource(DevidirGrupoExperimental, '/api/v1/dividirGrupoExperimental', endpoint='dividir_grupo_experimental')
+api.add_resource(DividirGrupoExperimental, '/api/v1/dividirGrupoExperimental', endpoint='dividir_grupo_experimental')
+api.add_resource(GrupoExperimental, '/api/v1/borrarGrupoExperimental/<int:idGrupoExperimental>', endpoint='borrar_grupo_experimental')
 
 #Jaula
 api.add_resource(Jaula, '/api/v1/nuevaJaula', endpoint="nueva_jaula")
@@ -124,7 +125,7 @@ api.add_resource(Animal, '/api/v1/animal/<int:idAnimal>', endpoint="animal")
 api.add_resource(Animal, '/api/v1/bajaAnimal/<int:idAnimal>', endpoint="baja_animal")
 api.add_resource(Animal, '/api/v1/nuevoAnimal', endpoint="nuevo_animal")
 api.add_resource(Animales, '/api/v1/animales', endpoint="animales")
-api.add_resource((Animales))
+#api.add_resource((Animales))
 api.add_resource(AnimalesSinJaula, '/api/v1/animalesSinJaula', endpoint="animales_sin_jaula")
 api.add_resource(AnimalesDeLaJaula, '/api/v1/jaula/<int:idJaula>/animales', endpoint="animales_de_la_jaula")
 api.add_resource(AnimalesDeLaJaula, '/api/v1/asignarJaulas', endpoint="asignar_jaulas")
