+++ resources/grupoExperimentalResource.py
@@ -13,14 +13,20 @@ class GrupoExperimental(Resource):
             return grupoExperimental, 200
         return {"Error" : "Se debe indicar el id del grupo experimental"}
 
+    def delete(self, idGrupoExperimental):
+        if idGrupoExperimental:
+            GrupoExperimentalService().borrarGrupoExperimental(idGrupoExperimental)
+            return  {"Status" : "Se borró el grupo experimental y sus subgrupos."}, 200
+        return {"Error" : "Se debe indicar el id del grupo experimental"}
+
     def post(self):
         datos = request.get_json()
         if datos:
             try:
                 GrupoExperimentalService().CrearGrupoExperimental(datos)
-                return {"Status": "OK"}, 200
-            except ValidationError as err:
-                return {"Error" : err.messages}, 400
+                return {"Status": "Se creó el grupo experimental"}, 200
+            except Exception as err:
+                return {"Error" : str(err)}, 400
         return {"Error" : "Se deben enviar datos para la creación de un grupo experimental"}, 400
 
 class GruposExperimentales(Resource):
@@ -34,7 +40,7 @@ class GruposExperimentales(Resource):
                 return {f"Status" : "No se encontraron grupos experimentales para el experimento {idExperimento}"}, 204
         return {"Error" : "Se debe enviar un id del experimento"}, 400
 
-class DevidirGrupoExperimental(Resource):
+class DividirGrupoExperimental(Resource):
 
     def post(self):
         datos = request.get_json()
