+++ servicios/animalService.py
@@ -1,6 +1,11 @@
-from models.mongo.fuenteExperimental import AnimalSchema, FuenteExperimental, FuenteExperimentalSchema, NuevoAnimalSchema
+from models.mongo.fuenteExperimental import FuenteExperimental
+
+from schemas.fuenteExperimentalSchema import FuenteExperimentalSchema
+                                             
 from models.mongo.jaula import Jaula
+from schemas.animalSchema import AnimalSchema,NuevoAnimalSchema
 from .validationService import Validacion
+
 
 class AnimalService:
 
@@ -21,21 +26,20 @@ class AnimalService:
     
     @classmethod
     def asignarJaulaAAnimales(cls, datos):
-        animales = AnimalSchema().load(datos, many=True)
-        print(animales)
-        cls.validarJaula(animales[0].id_jaula)
+        animales = AsignarAnimalAJaula().load(datos, many=True)
+        #cls.validarJaula(animales[0].id_jaula)
         for animal in animales:
+            cls.validarJaula(animal.id_jaula)
             FuenteExperimental.objects(id_fuenteExperimental =  animal.id_fuenteExperimental).update(id_jaula = animal.id_jaula)
     
     def validarJaula(idJaula):
         if not Validacion.existeLaJaulas(idJaula):
             raise Exception(f"No existe la jaula con id {idJaula} o no se encuentra habilitada.")
 
-
     @classmethod
     def todosLosAnimales(cls):
-        return AnimalSchema().dump(FuenteExperimental.objects(codigoGrupoExperimental__ne="", tipo ="Animal", baja=False).all(), many=True)
-    
+        return FuenteExperimental.objects().all()
+        #TO-DO PREGUNTAR: codigoGrupoExperimental__ne="", tipo ="Animal", baja=False)
     @classmethod
     def animalesSinJaula(cls):
         return AnimalSchema().dump(FuenteExperimental.objects(id_jaula = 0, tipo="Animal", codigoGrupoExperimental__ne="", baja=False).all(), many=True)
@@ -60,7 +64,10 @@ class AnimalService:
     
     @classmethod
     def animalesDelProyecto(cls, idProyecto):
-        animales = FuenteExperimental.objects(id_proyecto = idProyecto, codigoGrupoExperimental__ne="", tipo ="Animal", baja=False).all()
-        if animales:
-            return AnimalSchema().dump(animales, many=True)
-        return None
+        return FuenteExperimental.objects(id_proyecto = idProyecto, codigoGrupoExperimental__ne="", tipo ="Animal", baja=False).all()
+       
+    @classmethod
+    def actualizarProyectoAnimalesDeJaulas(cls,jaula):
+        FuenteExperimental.objects(id_jaula = jaula.id_jaula).update(set__id_proyecto =jaula.id_proyecto)
+
+
